requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-mptt==0.14.0
pillow==10.3.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
djangorestframework==3.15.2